changed  ushort to uint

+++ frame_buffer.c
@@ -1,17 +1,17 @@
 #include "frame_buffer.h"
 #include "io.h"
 
-void frame_buffer_write(ushort index, char character, char fg, char bg) 
+void frame_buffer_write(uint index, char character, char fg, char bg) 
 {
     char *frame_buffer_ptr = (char *) 0x000B8000;
     frame_buffer_ptr[index << 1] = character;
     frame_buffer_ptr[(index << 1) | 1] = ((bg & 0xF) << 4) | (fg & 0xF);
 }
 
-void frame_buffer_write_str(const char *buffer, ushort length) 
+void frame_buffer_write_str(const char *buffer, uint length) 
 {
-    ushort cursor_pos = frame_buffer_get_cursor_position();
-    ushort i = 0;
+    uint cursor_pos = frame_buffer_get_cursor_position();
+    uint i = 0;
     for (; i < length && buffer[i]; ++i) {
         frame_buffer_write(cursor_pos + i, buffer[i], 
             FRAME_BUFFER_COLOR_LIGHT_GREEN, FRAME_BUFFER_COLOR_BLACK);
@@ -24,7 +24,7 @@ void frame_buffer_write_str(const char *buffer, ushort length)
  *
  *  @param pos The new position of the cursor
  */
-void frame_buffer_move_cursor(ushort pos)
+void frame_buffer_move_cursor(uint pos)
 {
     outb(FRAME_BUFFER_COMMAND_PORT, FRAME_BUFFER_HIGH_BYTE_COMMAND);
     outb(FRAME_BUFFER_DATA_PORT,    ((pos >> 8) & 0x00FF));
@@ -35,9 +35,9 @@ void frame_buffer_move_cursor(ushort pos)
 /** frame_buffer_get_cursor_position:
  *  Return position of the cursor of the framebuffer.
  */
-ushort frame_buffer_get_cursor_position()
+uint frame_buffer_get_cursor_position()
 {
-    ushort pos = 0;
+    uint pos = 0;
     outb(FRAME_BUFFER_COMMAND_PORT, FRAME_BUFFER_HIGH_BYTE_COMMAND);
     pos = inb(FRAME_BUFFER_DATA_PORT);
     outb(FRAME_BUFFER_COMMAND_PORT, FRAME_BUFFER_LOW_BYTE_COMMAND);
